requirements.txt
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.44.0
accelerate>=0.33.0
deepspeed>=0.14.0 ; platform_system != "Windows"
peft>=0.12.0
datasets>=2.20.0
trl>=0.10.0
safetensors>=0.4.3
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.0.0
opencv-python>=4.10.0
pyyaml>=6.0.1
tqdm>=4.66.0
nltk>=3.9
pytest>=8.0.0
